:bug: fix: update app to new rest countries version

+++ app/src/main/java/com/example/countrylist/util/NumberFormatter.kt
@@ -21,7 +21,7 @@ object NumberFormatter {
         }
     }
 
-    fun toLocaleString(value: Int): String {
+    fun toLocaleString(value: Long): String {
         return String.format(Locale.US, "%,d", value)
     }
 }
